feat: prepare for new history plugin, remove podman (for now)
