--- container/process-compose.yaml
@@ -1,45 +1,0 @@
-version: "0.5"
-
-# Global environment is imported from project .env: $PROJECT/.env
-# in container, it's /home/user/tangram/.env
-
-log_location: /tmp/pc.log
-log_level: debug
-
-processes:
-  web:
-    working_dir: /home/user/tangram/web
-    command: just tangram-web
-    availability:
-      restart: "always"
-
-  channel:
-    command: /home/user/.cargo/bin/channel --port 2347 --jwt-secret secret --jwt-expiration-secs 315360000 # 10 years
-    working_dir: /home/user/tangram
-
-  core/restapi:
-    environment:
-    command: just tangram-restapi
-    working_dir: /home/user/tangram/
-    availability:
-      restart: "always"
-
-  core/planes:
-    working_dir: /home/user/tangram/
-    # command: env && pwd && uv run python -m tangram.planes
-    command: env && /usr/bin/planes --expire 600 --jet1090-channel jet1090
-    availability:
-      restart: "always"
-
-  # TODO merge trajectory and history_redis into one process
-  core/trajectory:
-    working_dir: /home/user/tangram/
-    command: env && pwd && uv run python -m tangram.trajectory
-    availability:
-      restart: "always"
-
-  core/history_redis:
-    working_dir: /home/user/tangram/
-    command: env && pwd && uv run python -m tangram.history_redis
-    availability:
-      restart: "always"
