+++ packages/tangram_history/pyproject.toml
@@ -1,0 +1,17 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "tangram_history"
+version = "0.2.0"
+dependencies = [
+    "pandas>=2",
+    "tangram>=0.2.0",
+]
+
+[project.entry-points."tangram.plugins"]
+history = "tangram_history:register_plugin"
+
+[tool.uv.sources]
+tangram = { workspace = true }
