+++ scripts/tangram.toml
@@ -2,6 +2,7 @@
 plugins = [
     "tangram_example",
     "tangram_jet1090",
+    "tangram_history",
     "tangram_system",
     "tangram_weather"
 ]
